Fix Outdated SDLActivity
While updating the SDL2 recipe to version 2.0.4 in PR #713, the
SDLActivity was not updated accordingly.

For example, SDL 2.0.4 expects the Activity to have a method called
"openAPKExpansionInputStream" while this method was called
"openAPKExtensionInputStream" in older versions. SDL2 is not able to
read files when the method is missing.

Also, text input is partly broken as it is not possible to enter a
"space" character.